Check annotation regrouping validity for abundance data in a SE object

This function checks the validity of regrouping the gene family/pathway abundance data to a requested target annotation. It ensures that the IDs in the abundance assay data from the SE object are known and belong to a valid annotation type. Then, it checks that only UniRef90 IDs are present for regrouping. Finally, it validates that the target annotation type is among the supported annotations.

Usage

.check_regroup(se, to)

Arguments

se

A SummarizedExperiment object containing gene/pathway abundance values as the assay and metadata as the column data.

to

A character string specifying the target annotation type for regrouping. Valid values include "ec" (Enzyme Comission), "eggnog" (evolutionary genealogy of genes: Non-supervised Orthologous Groups), "go" (Gene Ontology), "ko" (KEGG Orthology) and "pfam" (Protein families database of alignments and HMMs).

Value

'NULL' if all checks pass. If any issue is found, an error is raised.

Details

If any of these checks fail, an appropriate error message is displayed, and the function aborts the regrouping process.