Regroup HUMAnN3 Abundance Data in a SummarizedExperiment Object

This function regroups gene family or metabolic pathway abundance data stored in a SummarizedExperiment object and annotated with UniRef90 IDs (https://www.uniprot.org/help/uniref).

humann_regroup(se, to, version = "v201901b")

Arguments

se

A SummarizedExperiment object containing gene/pathway abundance values as the assay and metadata as the column data.

to

A character string specifying the target annotation type for regrouping. Valid values include "ec" (Enzyme Comission), "eggnog" (evolutionary genealogy of genes: Non-supervised Orthologous Groups), "go" (Gene Ontology), "ko" (KEGG Orthology) and "pfam" (Protein families database of alignments and HMMs).

version

A character string specifying the version of the database to use. Default is "v201901b".

Value

A SummarizedExperiment object containing gene/pathway abundance values regrouped according to the specified annotation as the assay, metadata as the column data and information about the regrouped functions as the row data.

Details

It takes as inputs a SummarizedExperiment object, the new annotation method to regroup to and the version of the database to use. First, the validity of inputs is checked using the .check_regroup() function. Then, it processes the abundance data, extracting its relevant UniRef90 IDs, and ensures that the necessary database file for the annotation conversion are available, fetching them if required. The function reads those files and matches them with abundance data UniRef90 annotation. Finally, it merges and aggregates the data based on the new annotations and constructs a new SummarizedExperiment object.

Examples

# Read HUMAnN3 & Regroup to PFAM annotation
se_regrouped <- read_humann(
  file_path = system.file("extdata", "reduced_genefam_rpk_uniref.tsv", package = "microfunk"),
  metadata = system.file("extdata", "reduced_meta.csv", package = "microfunk")
  ) %>%
  humann_regroup(to = "pfam")
#> ✔ File pfam_v201901b.tsv.gz cached