Read HUMAnN3 output file and create SummarizedExperiment objects

This function reads a HUMAnN3 output file (https://huttenhower.sph.harvard.edu/humann), which contains information about the abundance of gene families or metabolic pathways of a given sample in a TSV format. The function performs several checks on the structure and content of the file, as well as reading and validating the metadata CSV file, ensuring it contains the required information and matching sample IDs to the provided TSV data file. Finally, it creates a SummarizedExperiment object (https://bioconductor.org/packages/devel/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html) containing gene/pathway abundance values as the assay and metadata as the column data.

read_humann(file_path, metadata)

Arguments

file_path

Path to the HUMAnN3 output (gene families or pathway) collapsed TSV file. The file should contain aggregated abundance data across multiple samples. Each row represents a unique functional category (gene family / pathway) and each column corresponds to a sample.

metadata

Path to the CSV file containing sample IDs and additional information, where each row corresponds to a sample and each column represents a specific annotation associated with that sample.

Value

A SummarizedExperiment object containing gene/pathway abundance values as the assay and metadata as the column data.

Examples

# Def data paths
metadata <-
  system.file("extdata", "ex_meta.csv", package = "microfunk")
file_path <-
  system.file("extdata", "All_genefam_cpm_kegg.tsv", package = "microfunk")

# Read HUMAnN3
read_humann(file_path, metadata)
#> class: SummarizedExperiment 
#> dim: 81633 214 
#> metadata(0):
#> assays(1): humann
#> rownames(81633): K00001
#>   K00001|g__Acidaminococcus.s__Acidaminococcus_fermentans ...
#>   UNGROUPED|unclassified UNMAPPED
#> rowData names(0):
#> colnames(214): RAG_1 RAG_4 ... RAG_48 RAG_20
#> colData names(1): ARM