Run MaAsLin2 Analysis Quietly

This function performs a MaAsLin2 analysis (https://huttenhower.sph.harvard.edu/maaslin/), which finds associations between microbiome meta-omics features and metadata. Features consist on the abundance of gene families or metabolic pathways of multiple samples, obtained using HUMAnN3 (https://huttenhower.sph.harvard.edu/humann).

run_maaslin2(
  se,
  output = paste0(tempdir(), "/output_folder"),
  min_abundance = 0,
  min_prevalence = 0.1,
  min_variance = 0,
  normalization = "NONE",
  transform = "NONE",
  analysis_method = "LM",
  max_significance = 0.25,
  random_effects = NULL,
  fixed_effects = NULL,
  correction = "BH",
  standardize = TRUE,
  plot_heatmap = TRUE,
  heatmap_first_n = 50,
  plot_scatter = TRUE,
  max_pngs = 10,
  save_scatter = FALSE,
  save_models = FALSE,
  reference = NULL
)

Arguments

se

A SummarizedExperiment object containing normalized abundance values as the assay and metadata as the column data.

output

Character string specifying the output directory path. Default is a tmp folder.

min_abundance

Minimum abundance for each feature. Default is 0.0.

min_prevalence

Minimum percent of samples for which a feature is detected at minimum abundance. Default is 0.1.

min_variance

Minimum variance threshold for features to be included. Default is 0.0.

normalization

Normalization method to apply. Default is 'NONE'.

transform

Transformation method to be used. Default is 'NONE'.

analysis_method

Analysis method to apply. Default is 'LM'.

max_significance

Q-value threshold for significance. Default is 0.25.

random_effects

Random effects to be included in the analysis (comma-delimited for multiple effects). Default is NULL.

fixed_effects

Fixed effects to be included in the analysis (comma-delimited for multiple effects). Default is NULL.

correction

Correction method for computing the Q-value. Default is 'BH'.

standardize

Apply Z-score so continuous metadata are on the same scale. Default is FALSE.

plot_heatmap

Generate a heatmap for significant associations. Default is TRUE.

heatmap_first_n

In the heatmap, plot only top N features with significant associations. Default is 50.

plot_scatter

Generate scatter plots for the significant associations. Default is TRUE.

max_pngs

Set the maximum number of scatter plots for significant associations to save as png files. Default is 10.

save_scatter

Save all scatter plot ggplot objects to an RData file. Default is FALSE.

save_models

Return the full model outputs and save to an RData file. Default is FALSE.

reference

Factor to use as reference for categorical variables, provided as a string of 'variable,reference' (semi-colon delimited for multiple variables).

Value

Results of the MaAsLin2 analysis as a tibble.

Details

The function takes a SummarizedExperiment object containing normalized gene/ pathway abundance values as the assay and metadata as the column data, as well as several parameter options to run the analysis. It creates a temporary output folder to store created files and returns the result of the analysis on the console.

Examples

# Read HUMAnN3 & MaAsLin2 Analysis
da_result <- read_humann(
  file_path = system.file("extdata", "reduced_genefam_cpm_kegg.tsv", package = "microfunk"),
  metadata = system.file("extdata", "reduced_meta.csv", package = "microfunk")
) %>% run_maaslin2(fixed_effects = "ARM")

da_result
#> # A tibble: 3,806 × 11
#>    feature metadata da_method coefficient log2FC stderr lfc_stderr p_value
#>    <chr>   <chr>    <chr>           <dbl> <chr>   <dbl> <chr>        <dbl>
#>  1 K03592  ARM      maaslin2       -0.389 -           0 -                0
#>  2 K03788  ARM      maaslin2       -0.125 -           0 -                0
#>  3 K05708  ARM      maaslin2       -0.189 -           0 -                0
#>  4 K07270  ARM      maaslin2       -0.567 -           0 -                0
#>  5 K08678  ARM      maaslin2       -0.793 -           0 -                0
#>  6 K13652  ARM      maaslin2       -1.08  -           0 -                0
#>  7 K16950  ARM      maaslin2       -1.71  -           0 -                0
#>  8 K20861  ARM      maaslin2       -0.272 -           0 -                0
#>  9 K21055  ARM      maaslin2       -0.588 -           0 -                0
#> 10 K21065  ARM      maaslin2       -0.578 -           0 -                0
#> # ℹ 3,796 more rows
#> # ℹ 3 more variables: q_value <dbl>, padj_value <chr>, signif <lgl>