Fetch HUMAnN3 databases

This function fetches HUMAnN3 databases from AWS based on the provided annotation and version. If the databases do not exist in the cache directory, or if the overwrite option is enabled, the function will download the databases from AWS.

fetch_humann_db(
  annot = c("ec", "eggnog", "go", "ko", "pfam"),
  version = "v201901b",
  overwrite = FALSE
)

Arguments

annot

A character vector specifying the annotations to fetch. Default is c("ec", "eggnog", "go", "ko", "pfam").

version

A character string representing the version of the databases to fetch. Default is "v201901b".

overwrite

A logical indicating whether to overwrite existing files in the cache directory. Default is FALSE.

Value

invisible TRUE

Details

This function first checks if the provided annotation is valid. If not, it aborts with a message indicating the available annotations. Then, it creates the microfunk cache directory if it does not exist, and checks the version compatibility. Finally, it fetches the databases from AWS and saves them to the cache directory. If the overwrite option is TRUE, existing files in the cache directory will be overwritten.

Examples

# Fetch HUMAnN3 databases with default parameters
fetch_humann_db(annot = "test")
#> 
#> ── Fetching HUMAnN3 databases ──────────────────────────────────────────────────
#> ℹ Cache directory .microfunk already exists
#> ✔ File test_v201901b.tsv.gz cached